Finish surfaces before destroying.
This doesn't seem to be necessary for PNG output, but when trying to render to
PDF without this I only get an empty file. With flushing/finishing it works now.

Not sure if this is the right way to do it, perhaps there's actually a bug
in gi-cairo-render that prevents it from working without flush?